% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Features.R
\name{createFeatures}
\alias{createFeatures}
\title{createFeatures}
\usage{
createFeatures(x)
}
\arguments{
\item{x}{state}
}
\value{
Modified copy of state with metadata about features (covariates)
}
\description{
Create the features of included persons
}
\details{
Writes table to study database
}
\author{
William J. O'Brien
}
