% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Settings.R
\name{createInitialState}
\alias{createInitialState}
\title{createPredictionSettings}
\usage{
createInitialState(
  connectionDetails,
  cdmSchema,
  outputSchema,
  procedureStandardConcepts,
  procedureSourceConcepts,
  predictionInterval,
  continuousObservationInterval,
  controlToCaseRatio = NULL,
  covariateSettings = FeatureExtraction::createDefaultCovariateSettings(),
  andromedaFolder = "./",
  trainProportion = 0.8
)
}
\value{
List of class "ProcedurePrediction"
}
\description{
Create initial state
}
\details{
Returns a list of validated prediction settings
}
\author{
William J. O'Brien
}
