% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Labels.R
\name{createLabels}
\alias{createLabels}
\title{createLabels}
\usage{
createLabels(x)
}
\arguments{
\item{x}{state}
}
\value{
Modified copy of state with metadata about labels (outcomes)
}
\description{
Create the labels of included persons
}
\details{
Writes table to study database
}
\author{
William J. O'Brien
}
