% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Population.R
\name{createPopulation}
\alias{createPopulation}
\title{createInclusionCohort}
\usage{
createPopulation(x)
}
\arguments{
\item{x}{state}
}
\value{
Modified copy of state with metadata about study population
}
\description{
Create the cohort of included persons
}
\details{
Writes table to study database
}
\author{
William J. O'Brien
}
