% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Sampling.R
\name{createSample}
\alias{createSample}
\title{createSample}
\usage{
createSample(x)
}
\arguments{
\item{x}{state}
}
\value{
Modified copy of state with metadata about control cohort sample
}
\description{
Create a sample from the study population
}
\details{
Writes table to study database
}
\author{
William J. O'Brien
}
