% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Diagnostics.R
\name{createTestDiagnostics}
\alias{createTestDiagnostics}
\title{Diagnostics of test set predictions}
\usage{
createTestDiagnostics(x)
}
\arguments{
\item{x}{state}
}
\value{
Modified copy of state with information about test prediction diagnostics
}
\description{
Diagnostics of test set predictions
}
