% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Validation.R
\name{predictOnTestSet}
\alias{predictOnTestSet}
\title{Predict on data not used during training}
\usage{
predictOnTestSet(x)
}
\arguments{
\item{x}{state}
}
\value{
Modified copy of state with information about test performance in results list
}
\description{
Predict on data not used during training
}
