% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ModelFit.R
\name{trainModel}
\alias{trainModel}
\title{Fit regression model}
\usage{
trainModel(x)
}
\arguments{
\item{x}{state}
}
\value{
Modified copy of state with fit object in list of results
}
\description{
Fit regression model
}
