% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TrainTestSplit.R
\name{trainTestSplit}
\alias{trainTestSplit}
\title{Train/test split}
\usage{
trainTestSplit(x)
}
\arguments{
\item{x}{state}
}
\value{
Modified copy of state with metadata about train/test split
}
\description{
Writes new train/test tables to local andromeda objects
}
